Added 3 flags: (i) -all When used this flag allows the user to copy all genes from the $sourcedb instead of reading in a list
(ii) -remove_xrefs
When used this flag will remove xrefs from the copied genes before they are written to $outdb

(iii) -remove_stable_ids
When used this flag will remove stable_ids from the copied genes before they are written to $outdb

Renamed $db to $sourcedb so it is more clear which database genes are being copied from. Removed the -pass option for $sourcedb as we are only reading from it.