This is a simplified script which retained most of the functionalities except the handling of stable IDs, because we don't really map Ig stable IDs using this script anymore.
Other changes:

(1) Genes are fully loaded before being stored in the DB. This is to ensure
all the associated features/objects of a gene are captured and written
to the output DB.

(2) As genes are fully loaded, translations are loaded in the process, which
means both the query and target DBs need to have DNA, or they need to be
attached to a DNA_DB. This has been enforced.

(3) A full POD has been added.