Modified the if/else loop structure in _merge_redundant_transcript method:
(1) Now, for each Havana transcript, the code compares it to each Ensembl
transcript in the same locus and will execute the decision made for the
Ensembl transcript right after the comparison (i.e. merge, or not merge),
rather than loops through all Ensembl transcripts in the locus for one
Havana transcript and then executes only the decision made for the last
Ensembl transcript encountered.

(2) The code now correctly handles cases where are_matched_pair method
returns a value of 0, meaning the Ens and Hav transcripts compared differed
in their overall exon-intron structures, not only differed in their UTRs.

The code which deals with deletion of incomplete Ensembl transcripts is
very conservative as it stands now and needs tweaking in the future.