Fix deprecation warning for using defined($array) in perl 5.12.
For perl 5.12 onwards, defined(%alias) causes the following warning:
 defined(%hash) is deprecated at ...

The deprecation also holds true for defined(@alias). The correct
approach is to look into the typeglob to see if it has an ARRAY or
HASH entry. (This doesn't work for the SCALAR entry which always
exists if you check it, albeit as a reference to an undefined scalar.)