Fix blast parsing in Runnables, creates multiple runnables and refactorisation
The blast modules do not seem to use the blast params or not everywhere,
this will need some changes. Now it works for NCBI blast
Create multiple Infernal runnables instead of modifying the query on the fly,
this was causing problem when using File::Temp but it is also making the RunnableDB
behave correctly.
Goes into the working directory when running RNAFold to be able to delete the temporary
files properly