Module updates and new modules for cDNA update to work nice and easy
Now the unmapped objects are stored when we know we will not store them
like killlist just after they have been kill listed. The filter module
is generating most of the unmapped object, verbosity has to be set to 2.
New modules and their test, HiveDownloadData.pm aim is to do all file
fetching. Some of them are only for the cDNA update pipeline.