Added 2 option for best targetted to avoid creating poor quality models
Some models are kept even if they have low identity and coverage compared
to the original protein. As we are in species specific context and the
models would probably not be pseudogenes, it should be OK. The default is
the old behaviour.
The new options are protein_min_coverage and protein_min_identity