Set all genes from the cdna database to be cdna_alignment
We do some filtering in the cdna database for cdna2genome.
In some cases we do not want to have all the alignments in
the otherfeature database. Because we will not have much
cDNAs in the future it seems overkill to do it the old way.
Previously we would have not used the alignments and removed
all the alignments failing the thresholds