Download and load the real sequence when loading an assembly with toplevel only
For some reasons salmon scaffold had empty length which broke the module.
Download and check the genomic file.
Load the real sequence, we assume that the sequence in the FASTA file
  are in the same order as the report file
Do not flow the genomic file path and md5sum on the _branch_to_flow_to
  branch
Reorganised the config so it goes directly to load_meta_info