Bug fix: set start of start Exon correctly when adding 3' UTR mutliple times
In rare cases where we are trying to add 3' UTR, it fails to set the wrong start.
This is because we need to update the phase when adding 5' UTR as it should be -1.
If your phase was 1 or 2 we need to change the start so the translation start at
the correct position.