Read the IMGT EMBL file instead of a fasta file
The idea is to create the protein sequence and the mRNA sequence
for each item in the file. Thus allowing to use different parameters
based on the Ig class of the sequence. For example, an IgV will never
have an intron bigger that 1000bp and an IgJ will always be single
exon.
There is code to try to split sequences which have not been annotated
yet and where the different segments should be split.
More work need to be done on the parsing but using this module can help:
modules/Bio/EnsEMBL/Analysis/Tools/IMGT/SeqIO/imgt_embl.pm