New module to create directories
In some case we need to be able to share directories, this module
create the directories recursively and change the permissions
according to parameters given. The default is that user and group
can write to the directories.
If using LFS, it can also stripe the directories when files to
generate will be big.