Move the concat to file to a single file
The module was appending to a single file which is quite dangerous
when you can have multiple jobs running at the same time. It also
doesn't check that the file was properly closed. Now the module writes
in one file per job. That means a file concatenation is needed later in
the pipeline.