Remove force_drop from Genome config file
force_drop in the HiveCreateDatabases module should only be used
on the command line or through guiHive as it would try to delete
the database before creating it. Having the parameter set in the
config can cause problem as not everyone would be aware of it and
may accidentally delete some precious databases.