Remove gw_nogtag from LayerAnnotationStatic
Models generated by Genewise without forcing canonical splice sites
can be very wrong. We could have a higher threshold but that would
need a proper investigation.
I did not remove the analysis in the pipeline but it would be the
next logical step. Instead I removed the biotype from the layers
so the models do not pollute the gene set.