Add clean_utrs as method to be used in CleanUTRs and GeneBuilder
I moved the entire algorithm from the CleanUTRs runnableDB to GeneUtils
so it can be used by multiple runnableDBs.
Currently the default values are the same in the method and both
runnableDBs. Either the values should be set to undef in default_params
of the runnableDBs and be changed in the library or it should be set the
same in all modules the same way it is now.