Do store in write_output instead of run
Because we are already loading all the models in order to speed up the
db queries, we are already using more memory that probably needed. So we
can also delete the genes in write_output instead of run.
It also has advantages that if your job dies for any reasons it should
have less chances to have broken genes.
It also helps for debugging as you can use -no_write