Add long reads to otherfeatures.
Refactored the pipeline so it is easier to add new databases like ONT or
BRAKER. New databases need to be added to 'databases_to_check', the
pipeline will look for the database existence and if there are genes.
Further analyses can be done after the copy happened by adding or
modifying a fan

Added an option to HiveCopyGenes to set the analysis and the biotype for
all the object copied

Removed the population of some Production tables as the Production
DBSync pipeline will need to be run in any case

Automatically add the analyses and analyses description to the
Production database. The script is using the JSON module for better code

Removed unused parameters from the main config