Delete merged genes. Add analyses to production DB
When we create the scallop gene set the merge data is not needed like it
was for the old pipeline. So the merged gene set can be deleted and it
will save space and compute time.
Add all the analyses to the production DB. If someone need to do an
update they can use the -update flag on the script