Parse minimap2 CIGAR string to create transcripts
Parse the CIGAR string to generate the transcripts. A deletetion will be
modeled with an intron. The CIGAR will be converted as an "Ensembl"
cigar string where an insertion is equivalent to a deletion. A softmask
operation will remove sequence from the cDNA. Adn the CIGAR string is
reverted when the alignmednt was on the reverse strand
Identity and coverage is calculated using the cs flag.
It expects either a Bio::EnsEMBL::Slice as query or a hash of
Bio::EnsEMBL::Slice to be able to assign Slice to the models.