Add fasta DNA fetching
The new module FastaSequenceAdaptor.pm inherits from the SequenceAdaptor
object but fetches sequences from a fasta file stored in the method fasta.
It needs some hacking of the DBAdaptor but once this is done any
DBAdaptor can use its own fasta to retrieve DNA
The config for the pipeline is updated to use the correct options. The
source fasta must be dumped/downloaded and indexed before starting the
pipeline.