Use CIGAR string from minimap to generate transcripts
Use the CIGAR string from minimap2 to generate the models. The original
CIGAR is stored as a transcript supporting feature. The exon CIGAR is
trasformed as an Ensembl CIGAR:
  I becomes D
  D becomes an intron
  S and H are not added but impact the hstart and hend
  X and = becomes M
Added a method to generate the CDS from the CIGAR string and the source
transcript
Transcript is transfered to toplevel when method is minimap_local