Refactor memory management. Use sub_Slice
The current memory branching was hard coded. The user can now provide a
list of memory limits to check and if the config is updated to redirect
to the correct branch.
Use sub_Slice instead of SliceAdaptor->fetch to reduce possible calls to
the database
Fix a bug where the start of the slice could be 0