Repeatmask pipeline set meta_key 'repeat.analysis'
The pipeline set the meta keys for dust and trf if there is at least one
row from these analyses in the repeat_feature table.
Three new parameters are added first_choice_repeat, second_choice_repeat
and third_choice_repeat which should contains the logic_name of an
analysis. They are tested in order and the first one with data in the
repeat_feature table is used. If the force red parameters is used in the
ini file then red is set as first_choice_repeat.
DumpGenome is checking the meta table for 'repeat.analysis' if is there
is no repeat_logic_names specified and if mask is set to 1