Extend translation if phase > 0 and has UTR
In some case we would align a transcript but we would be missing the
start of the translation. So the phase of the first coding exon would be
superior to 0 but we would have UTRs. This would confuse the pep2genomic
method which is used when removing internal stop codon and break
everything.